Add review comments
For the ER diagram, see https://www.lucidchart.com/pages/er-diagrams
The current ER diagram, while displays the entities and their fields, does not look like a proper ER diagram.

For the architecture diagram - while the components of the system are shown, the interactions between them must be the protocols or APIs they communicate with, not the semantics of the actual work being done (e.g. UI -- HTTP --> Server and not UI -- Initiates session --> <-- HTTP response -- Server)

Binaries must not be included in repos.

The Readme is empty

Missing explicit dependency versions in pom.xml

"com.example.demo" is not an appropriate package for this project

There is a mixing of RPC and REST-style endpoint paths

Some DTOs are missing equals() and hashCode()

The entire "cart" package contains only commented out code...